refactor(extraction): split oversized modules and remove stale files (#147)
- focus_optimizer.py (1767→646 lines): split into field_cache.py (SlabLRUCache + FieldCache) and hotspot_scoring.py (air-point scoring functions)
- field_combiner.py (592→426 lines): sliced combination moved to field_combiner_sliced.py; re-exported for backward compat
- auto_induced_processor.py (868→511 lines): extraction methods moved to _AutoInducedExtractionMixin in auto_induced_extractors.py
- delete tests/test_s4l_multiprocessing_extended.py and tests/test_s4l_import_order.py
- clean up cloud_setup/license_automation.py: no global mutation, _click_button() helper, exceptions instead of None tuples

Closes #147
