+++ goliat/extraction/auto_induced_processor.py
@@ -15,7 +15,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 from ..logging_manager import LoggingMixin
-from .field_combiner import combine_fields_sliced
+from .auto_induced_extractors import _AutoInducedExtractionMixin
+from .field_combiner_sliced import combine_fields_sliced
 from .focus_optimizer import find_focus_and_compute_weights
 
 if TYPE_CHECKING:
@@ -38,7 +39,7 @@ class _CombineRequest:
     field_caches: dict | None = None
 
 
-class AutoInducedProcessor(LoggingMixin):
+class AutoInducedProcessor(LoggingMixin, _AutoInducedExtractionMixin):
     """Processes auto-induced exposure using existing SAPD extraction infrastructure.
 
     Orchestrates focus search, field combination, and SAPD extraction for
@@ -115,7 +116,7 @@ class AutoInducedProcessor(LoggingMixin):
         with self.study.subtask("auto_induced_combine_fields"):
             combined_h5_paths = []
             from tqdm import tqdm
-            from .focus_optimizer import FieldCache
+            from .field_cache import FieldCache
 
             # Pre-load all source fields once, shared across all candidates.
             # FieldCache auto-detects available RAM:
@@ -487,363 +488,6 @@ class AutoInducedProcessor(LoggingMixin):
         except Exception as e:
             self._log(f"      ERROR combining fields: {e}", log_type="error")
             return None
-
-    def _extract_sar(
-        self,
-        combined_h5: Path,
-        candidate_idx: int,
-        candidate: dict,
-        input_h5: Path,
-        candidate_output_dir: Path | None = None,
-        save_intermediate_files: bool = False,
-    ) -> dict:
-        """Run the full SAR extraction pipeline on a combined H5 file.
-
-        Uses the existing SarExtractor and Reporter classes (via AutoInducedSarContext shim)
-        to produce the same outputs as a normal simulation extraction:
-          - sar_results.json  (full metrics: whole-body SAR, peak 10g, group SAR, peak details)
-          - sar_stats_all_tissues.pkl  (tissue DataFrame + groups for programmatic access)
-          - sar_stats_all_tissues.html  (human-readable tissue table)
-
-        These are written to candidate_output_dir (e.g. auto_induced/candidate_01/).
-
-        The project must already be open with phantom geometry loaded so that
-        the SAR evaluator can map E-field to tissue conductivity.
-
-        Args:
-            combined_h5: Path to combined _Output.h5 file (full-volume, E-field only).
-            candidate_idx: 1-based candidate index (used for naming and logging).
-            candidate: Candidate dict (not used for SAR, kept for API consistency).
-            input_h5: Path to an _Input.h5 file (not used for SAR, kept for API consistency).
-            candidate_output_dir: Directory to write per-candidate results into.
-                Defaults to combined_h5.parent / f"candidate_{candidate_idx:02d}".
-            save_intermediate_files: If True, save the Sim4Life project after extraction
-                for debugging.
-
-        Returns:
-            Dict with at minimum 'candidate_idx' and 'peak_sar_10g_W_kg'. On error,
-            contains 'error' key instead of metric values.
-        """
-        from .auto_induced_sar_context import AutoInducedSarContext, AutoInducedReporter, save_candidate_json
-        from .sar_extractor import SarExtractor
-
-        if candidate_output_dir is None:
-            candidate_output_dir = combined_h5.parent / f"candidate_{candidate_idx:02d}"
-        candidate_output_dir.mkdir(parents=True, exist_ok=True)
-
-        self.verbose_logger.info(f"Extracting SAR (full pipeline) from {combined_h5.name} -> {candidate_output_dir.name}/")
-
-        try:
-            import s4l_v1.analysis as analysis
-            import s4l_v1.document as document
-
-            # --- Step 1: Create SimulationExtractor from the combined H5 ---
-            sim_extractor = analysis.extractors.SimulationExtractor(inputs=[])
-            sim_extractor.Name = f"AutoInduced_SAR_Candidate{candidate_idx}"
-            sim_extractor.FileName = str(combined_h5)
-            sim_extractor.UpdateAttributes()
-            document.AllAlgorithms.Add(sim_extractor)
-
-            # --- Step 2: Build context shim and run SarExtractor ---
-            context = AutoInducedSarContext(self, candidate_output_dir)
-            results_data: dict = {}
-
-            sar_extractor = SarExtractor(context, results_data)  # type: ignore[arg-type]
-            sar_extractor.extract_sar_statistics(sim_extractor)
-
-            # --- Step 3: Save reports if extraction succeeded ---
-            if "_temp_sar_df" in results_data:
-                reporter = AutoInducedReporter(context)  # type: ignore[arg-type]
-                reporter.save_reports(
-                    results_data.pop("_temp_sar_df"),
-                    results_data.pop("_temp_tissue_groups"),
-                    results_data.pop("_temp_group_sar_stats"),
-                    results_data,
-                )
-                save_candidate_json(results_data, candidate_output_dir)
-            else:
-                self._log(
-                    f"      Candidate #{candidate_idx}: SarExtractor produced no data (check logs)",
-                    log_type="warning",
-                )
-
-            # --- Step 4: Cleanup Sim4Life algorithms ---
-            try:
-                document.AllAlgorithms.Remove(sim_extractor)
-            except Exception:
-                pass
-
-            # Optional: save debug project
-            if save_intermediate_files:
-                try:
-                    debug_path = str(combined_h5).replace("_Output.h5", "_intermediate.smash")
-                    document.SaveAs(debug_path)
-                except Exception:
-                    pass
-
-            # Build return dict for worst-case comparison and summary JSON
-            result = {
-                "candidate_idx": candidate_idx,
-                "peak_sar_10g_W_kg": results_data.get("peak_sar_10g_W_kg"),
-                "whole_body_sar_W_kg": results_data.get("whole_body_sar"),
-                "combined_h5": str(combined_h5),
-                "output_dir": str(candidate_output_dir),
-            }
-
-            peak = result["peak_sar_10g_W_kg"]
-            self._log(
-                f"      Candidate #{candidate_idx}: peak SAR 10g = {peak:.4e} W/kg"
-                if peak is not None
-                else f"      Candidate #{candidate_idx}: SAR extraction produced no peak value",
-                log_type="info",
-            )
-
-            return result
-
-        except Exception as e:
-            self._log(f"      ERROR extracting SAR for candidate #{candidate_idx}: {e}", log_type="error")
-            import traceback
-
-            self.verbose_logger.error(traceback.format_exc())
-            return {
-                "candidate_idx": candidate_idx,
-                "error": str(e),
-            }
-
-    def _read_grid_axes(self, input_h5: Path) -> "tuple | None":
-        """Read grid axis arrays from input H5, returning (x, y, z) in mm, or None."""
-        import h5py
-
-        with h5py.File(input_h5, "r") as f:
-            for mesh_key in f["Meshes"].keys():
-                mesh = f[f"Meshes/{mesh_key}"]
-                if "axis_x" in mesh:
-                    return (
-                        mesh["axis_x"][:] * 1000,
-                        mesh["axis_y"][:] * 1000,
-                        mesh["axis_z"][:] * 1000,
-                    )
-        return None
-
-    def _resolve_skin_entity(self, candidate_idx: int, model) -> Any:
-        """Return a united skin entity from cache or by uniting model entities.
-
-        Returns None if no skin entities are found.
-        """
-        cache_dir = os.path.join(self.config.base_dir, "data", "phantom_skins")
-        cache_path = os.path.join(cache_dir, f"{self.phantom_name}_skin.sab")
-        united_entity = None
-
-        if os.path.exists(cache_path):
-            try:
-                imported_entities = list(model.Import(cache_path))
-                if imported_entities:
-                    united_entity = imported_entities[0]
-                    united_entity.Name = f"AutoInduced_CachedSkin_{candidate_idx}"
-            except Exception:
-                pass
-
-        if united_entity is None:
-            skin_entity_names = self._get_skin_entity_names()
-            all_entities = model.AllEntities()
-            skin_entities = [all_entities[n] for n in skin_entity_names if n in all_entities]
-
-            if not skin_entities:
-                return None
-
-            united_entity = model.Unite([e.Clone() for e in skin_entities]) if len(skin_entities) > 1 else skin_entities[0].Clone()
-
-            try:
-                os.makedirs(cache_dir, exist_ok=True)
-                model.Export([united_entity], cache_path)
-            except Exception:
-                pass
-
-        return united_entity
-
-    @staticmethod
-    def _parse_sapd_data_collection(data_collection) -> "tuple[Any, Any]":
-        """Extract peak SAPD value and location from a DataSimpleDataCollection."""
-
-        def _safe_get(key):
-            try:
-                return data_collection.FieldValue(key, 0)
-            except TypeError:
-                return None
-
-        keys = list(data_collection.Keys())
-        peak_sapd = peak_loc = None
-
-        for key in keys:
-            val = _safe_get(key)
-            if val is not None:
-                if "Peak" in key and "Power" in key:
-                    peak_sapd = val
-                if "Peak" in key and "Location" in key:
-                    peak_loc = val
-
-        if peak_sapd is None:
-            for key in keys:
-                if "Power" in key or "Density" in key:
-                    val = _safe_get(key)
-                    if val is not None:
-                        peak_sapd = val
-                        break
-
-        return peak_sapd, peak_loc
-
-    def _extract_sapd(
-        self,
-        combined_h5: Path,
-        candidate_idx: int,
-        candidate: dict,
-        cube_size_mm: float,
-        input_h5: Path,
-        save_intermediate_files: bool = False,
-    ) -> dict:
-        """Extract SAPD from a combined H5 file.
-
-        Uses the existing SAPD extraction infrastructure. The project should
-        already be open with the phantom geometry loaded.
-
-        Args:
-            combined_h5: Path to combined _Output.h5 file.
-            candidate_idx: 1-based candidate index.
-            candidate: Candidate dict with 'voxel_idx' for center location.
-            cube_size_mm: Size of the bounding box for mesh slicing.
-            input_h5: Path to an input H5 to read grid axes for voxel->mm conversion.
-            save_intermediate_files: If True, save smash file after extraction for inspection.
-
-        Returns:
-            Dict with SAPD extraction results.
-        """
-        from ..utils.mesh_slicer import slice_entity_to_box, voxel_idx_to_mm
-
-        self.verbose_logger.info(f"Extracting SAPD from {combined_h5.name}")
-
-        try:
-            import XCoreModeling
-            import s4l_v1.analysis as analysis
-            import s4l_v1.document as document
-            import s4l_v1.model as model
-            import s4l_v1.units as units
-
-            # Phase 1: resolve grid axes and compute focus centre
-            axes = self._read_grid_axes(input_h5)
-            if axes is None:
-                return {"candidate_idx": candidate_idx, "error": "Could not read grid axes from input H5"}
-
-            center_mm = voxel_idx_to_mm(candidate["voxel_idx"], axes)
-            self.verbose_logger.info(f"  Focus center: {center_mm} mm")
-
-            # Phase 2: set up EM field extractor
-            sim_extractor = analysis.extractors.SimulationExtractor(inputs=[])
-            sim_extractor.Name = f"AutoInduced_Candidate{candidate_idx}"
-            sim_extractor.FileName = str(combined_h5)
-            sim_extractor.UpdateAttributes()
-            document.AllAlgorithms.Add(sim_extractor)
-
-            em_sensor_extractor = sim_extractor["Overall Field"]
-            em_sensor_extractor.FrequencySettings.ExtractedFrequency = "All"
-            em_sensor_extractor.UpdateAttributes()
-            document.AllAlgorithms.Add(em_sensor_extractor)
-            em_sensor_extractor.Update()
-
-            # Phase 3: resolve skin surface and slice to bounding cube
-            united_entity = self._resolve_skin_entity(candidate_idx, model)
-            if united_entity is None:
-                return {"candidate_idx": candidate_idx, "error": "No skin entities found in project"}
-
-            surface_entity = united_entity.Clone()
-            surface_entity.Name = f"AutoInduced_Skin_{candidate_idx}"
-            surface_entity, sliced = slice_entity_to_box(
-                entity=surface_entity,
-                center_mm=center_mm,
-                side_len_mm=cube_size_mm,
-                model_module=model,
-                xcoremodeling_module=XCoreModeling,
-                logger=self.verbose_logger,
-            )
-            if sliced:
-                surface_entity.Name = f"AutoInduced_Skin_Sliced_{candidate_idx}"
-            else:
-                self.verbose_logger.warning("  Mesh slicing failed, using full skin mesh")
-
-            # Phase 4: run SAPD evaluator
-            model_to_grid_filter = analysis.core.ModelToGridFilter(inputs=[])
-            model_to_grid_filter.Name = f"AutoInduced_SkinSurface_{candidate_idx}"
-            model_to_grid_filter.Entity = surface_entity
-            model_to_grid_filter.UpdateAttributes()
-            document.AllAlgorithms.Add(model_to_grid_filter)
-
-            inputs = [em_sensor_extractor.Outputs["S(x,y,z,f0)"], model_to_grid_filter.Outputs["Surface"]]
-            sapd_evaluator = analysis.em_evaluators.GenericSAPDEvaluator(inputs=inputs)
-            sapd_evaluator.AveragingArea = 4.0, units.SquareCentiMeters
-            sapd_evaluator.Threshold = 0.01, units.Meters
-            sapd_evaluator.UpdateAttributes()
-            document.AllAlgorithms.Add(sapd_evaluator)
-
-            sapd_report = sapd_evaluator.Outputs["Spatial-Averaged Power Density Report"]
-            sapd_report.Update()
-
-            # Phase 5: parse results
-            data_collection = sapd_report.Data.DataSimpleDataCollection
-            if not data_collection:
-                return {"candidate_idx": candidate_idx, "error": "No SAPD data available"}
-
-            peak_sapd, peak_loc = self._parse_sapd_data_collection(data_collection)
-
-            # Optional: save debug artefacts
-            if save_intermediate_files:
-                try:
-                    focus_point = model.CreatePoint(model.Vec3(center_mm[0], center_mm[1], center_mm[2]))
-                    focus_point.Name = f"Focus_Center_Candidate{candidate_idx}"
-                except Exception:
-                    pass
-                try:
-                    document.SaveAs(str(combined_h5).replace("_Output.h5", "_intermediate.smash"))
-                except Exception:
-                    pass
-
-            result = {
-                "candidate_idx": candidate_idx,
-                "peak_sapd_w_m2": float(peak_sapd) if peak_sapd else None,
-                "peak_location_m": list(peak_loc) if peak_loc else None,
-                "combined_h5": str(combined_h5),
-            }
-
-            # Phase 6: cleanup — remove temporary algorithms and entities
-            try:
-                document.AllAlgorithms.Remove(sapd_evaluator)
-                document.AllAlgorithms.Remove(model_to_grid_filter)
-                document.AllAlgorithms.Remove(em_sensor_extractor)
-                document.AllAlgorithms.Remove(sim_extractor)
-                surface_entity.Delete()
-                united_entity.Delete()
-            except Exception:
-                pass
-
-            return result
-
-        except Exception as e:
-            self._log(f"      ERROR extracting SAPD: {e}", log_type="error")
-            import traceback
-
-            self.verbose_logger.error(traceback.format_exc())
-            return {"candidate_idx": candidate_idx, "error": str(e)}
-
-    def _get_skin_entity_names(self) -> list[str]:
-        """Get skin entity names from config.
-
-        Returns:
-            List of skin entity names (e.g., ["Skin", "Ear_skin"]).
-        """
-        try:
-            material_mapping = self.config.get_material_mapping(self.phantom_name)
-            tissue_groups = material_mapping.get("_tissue_groups", {})
-            return tissue_groups.get("skin_group", ["Skin", "Ear_skin"])
-        except Exception:
-            return ["Skin", "Ear_skin"]
 
     def _find_worst_case(self, results: list[dict], metric_key: str = "peak_sapd_w_m2") -> dict:
         """Find the worst-case (highest metric) result.
